remove dupes and update call sites, move enum
