+++ src/types/sorting.ts
@@ -10,14 +10,6 @@ export interface ObservationSortAPIParams {
   order: SortDirection;
 }
 
-export enum OBSERVATIONS_SORT {
-  DATE_UPLOADED_NEWEST = "DATE_UPLOADED_NEWEST",
-  DATE_UPLOADED_OLDEST = "DATE_UPLOADED_OLDEST",
-  DATE_OBSERVED_NEWEST = "DATE_OBSERVED_NEWEST",
-  DATE_OBSERVED_OLDEST = "DATE_OBSERVED_OLDEST",
-  MOST_FAVED = "MOST_FAVED",
-}
-
 // eventually this will be a union type that accounts for different types of sorting, i.e. name
 export type SpeciesSortBy = "count"
